+++ package.json
@@ -19,10 +19,6 @@
       "types": "./dist/index.d.ts",
       "default": "./dist/index.mjs"
     },
-    "./brand": {
-      "types": "./dist/brand/index.d.ts",
-      "default": "./dist/brand/index.mjs"
-    },
     "./server-specification": {
       "types": "./dist/server-specification/index.d.ts",
       "default": "./dist/server-specification/index.mjs"
@@ -30,10 +26,6 @@
     "./storage": {
       "types": "./dist/storage/index.d.ts",
       "default": "./dist/storage/index.mjs"
-    },
-    "./types": {
-      "types": "./dist/types.d.ts",
-      "default": "./dist/types.mjs"
     }
   },
   "keywords": [
