+++ src/index.ts
@@ -1,4 +1,15 @@
+export {
+  createIdempotencyFingerprint,
+  createIdempotentStorageKey,
+  type IdempotencyFingerprint,
+  type IdempotentStorageKey,
+} from "./brand";
+
 export { IdempotencyKeyStorageError, UnsafeImplementationError } from "./error";
 
-export type { IdempotentRequestImplementation } from "./middleware";
-export { idempotentRequestUniversalMiddleware } from "./middleware";
+export {
+  type IdempotentRequestImplementation,
+  idempotentRequestUniversalMiddleware,
+} from "./middleware";
+
+export type { SerializedResponse } from "./utils/response";
